move printf logic in vprintf
