+++ kernel/include/stdio.h
@@ -3,4 +3,5 @@
 #include <stdarg.h>
 #include <stddef.h>
 
+int vprintf(const char* restrict format, va_list ap);
 int printf(const char* restrict format, ...);
